+++ packages/vrender-components/src/axis/overlap/auto-wrap.ts
@@ -1,5 +1,5 @@
 import type { IText } from '@visactor/vrender-core';
-import { isEmpty, isValidNumber } from '@visactor/vutils';
+import { isEmpty, isValidNumber, min } from '@visactor/vutils';
 import { isAngleHorizontal, isAngleVertical } from './util';
 
 type WrapConfig = {
@@ -15,25 +15,38 @@ export function autoWrap(labels: IText[], config: WrapConfig) {
     return;
   }
 
-  const verticalLimitLength = axisLength / labels.length;
+  // 注意：自动换行算法暂时只考虑所有标签角度都一致的情况
+  const firstLabel = labels[0];
+  const angle = firstLabel.attribute.angle;
+  const isHorizontal = isAngleHorizontal(angle);
+  const isVertical = isAngleVertical(angle);
+  const isX = orient === 'top' || orient === 'bottom';
 
-  labels.forEach(label => {
-    const angle = label.attribute.angle;
+  let verticalLimitLength = axisLength / labels.length;
 
-    const isHorizontal = isAngleHorizontal(angle);
-    const isVertical = isAngleVertical(angle);
-    const isX = orient === 'top' || orient === 'bottom';
-
+  labels.forEach((label, index) => {
     if (isX) {
       if (isVertical && Math.floor(label.AABBBounds.height()) <= limitLength) {
         return;
       }
-      if (isHorizontal && Math.floor(label.AABBBounds.width()) <= verticalLimitLength) {
-        return;
+      if (isHorizontal) {
+        const curLabelX = label.attribute.x;
+        const nextLabelX = labels[index + 1]?.attribute.x;
+        const lastLabelX = labels[index - 1]?.attribute.x;
+        const minGap = getLabelMinGap(curLabelX, nextLabelX, lastLabelX);
+        if (isValidNumber(minGap)) {
+          verticalLimitLength = min(verticalLimitLength, minGap);
+        }
       }
     } else {
-      if (isVertical && Math.floor(label.AABBBounds.height()) <= verticalLimitLength) {
-        return;
+      if (isVertical) {
+        const curLabelY = label.attribute.y;
+        const nextLabelY = labels[index + 1]?.attribute.y;
+        const lastLabelY = labels[index - 1]?.attribute.y;
+        const minGap = getLabelMinGap(curLabelY, nextLabelY, lastLabelY);
+        if (isValidNumber(minGap)) {
+          verticalLimitLength = min(verticalLimitLength, minGap);
+        }
       }
       if (isHorizontal && Math.floor(label.AABBBounds.width()) <= limitLength) {
         return;
@@ -60,7 +73,6 @@ export function autoWrap(labels: IText[], config: WrapConfig) {
         heightLimit = verticalLimitLength;
       }
     }
-
     label.setAttributes({
       maxLineWidth: limitLabelLength,
       ellipsis: label.attribute.ellipsis ?? ellipsis,
@@ -68,4 +80,21 @@ export function autoWrap(labels: IText[], config: WrapConfig) {
       heightLimit
     });
   });
+}
+
+function getLabelMinGap(current: number, next?: number, prev?: number) {
+  let minGap;
+  if (isValidNumber(next)) {
+    minGap = Math.abs(next - current);
+  }
+
+  if (isValidNumber(prev)) {
+    if (isValidNumber(minGap)) {
+      minGap = Math.min(Math.abs(current - prev), minGap);
+    } else {
+      minGap = Math.abs(current - prev);
+    }
+  }
+
+  return minGap;
 }