+++ packages/vrender-core/CHANGELOG.md
@@ -1,6 +1,44 @@
 # Change Log - @visactor/vrender-core
 
-This log was last generated on Tue, 15 Oct 2024 03:50:15 GMT and should not be manually modified.
+This log was last generated on Wed, 13 Nov 2024 07:47:16 GMT and should not be manually modified.
+
+## 0.20.14
+Wed, 13 Nov 2024 07:47:16 GMT
+
+_Version update only_
+
+## 0.20.13
+Wed, 13 Nov 2024 06:35:02 GMT
+
+### Updates
+
+- feat: support auto refresh plugin
+- fix: fix smartInvert of gradient bar
+
+
+- fix: fix issue with text clip
+
+## 0.20.12
+Thu, 31 Oct 2024 02:49:49 GMT
+
+_Version update only_
+
+## 0.20.11
+Wed, 30 Oct 2024 13:10:03 GMT
+
+### Updates
+
+- fix: fix group fill gradient #1518
+- fix: annotate the Ramer Douglas Peucker algorithm in point optimization
+
+## 0.20.10
+Wed, 23 Oct 2024 08:37:33 GMT
+
+### Updates
+
+- feat: support fillStrokeOrder, closed #1505
+- fix: fix issue with parse m where multi pos follow, closed #1490
+- fix: fix the accuracy issue of number matching, closed #1488
 
 ## 0.20.9
 Tue, 15 Oct 2024 03:50:15 GMT